model finito (per ora)

+++ model/datiModelli.cpp
@@ -1,23 +1,23 @@
 #include "datiModelli.h"
 
-datiModelli::datiModelli(QString n, QString t, double tu, double ctm, double cb, double cl,double cv, unsigned int pg):
+datiModelli::datiModelli(QString n, QString t, float tu, float ctm, float cb, float cl,float cv, unsigned int pg):
     nomeModello(n), tessuto(t), tessUsato(tu), costoTessutoMq(ctm), costoBase(cb), costoLavaggio(cl), costoVendita(cv), produzioneGiornaliera(pg) {}
 
 const QString& datiModelli::getNomeModello() const{return nomeModello;}
 
 const QString& datiModelli::getTessuto() const{return tessuto;}
 
-double datiModelli::getTessUsato() const{return tessUsato;}
+float datiModelli::getTessUsato() const{return tessUsato;}
 
-double datiModelli::getCostoTessutoMq() const{return costoTessutoMq;}
+float datiModelli::getCostoTessutoMq() const{return costoTessutoMq;}
 
-double datiModelli::getCostoBase() const{return costoBase;}
+float datiModelli::getCostoBase() const{return costoBase;}
 
-double datiModelli::getCostoLavaggio() const{return costoLavaggio;}
+float datiModelli::getCostoLavaggio() const{return costoLavaggio;}
 
-double datiModelli::getCostoFinale() const{return tessUsato*costoTessutoMq+costoBase+costoLavaggio;}
+float datiModelli::getCostoFinale() const{return tessUsato*costoTessutoMq+costoBase+costoLavaggio;}
 
-double datiModelli::getCostoVendita() const{return costoVendita;}
+float datiModelli::getCostoVendita() const{return costoVendita;}
 
 unsigned int datiModelli::getProduzioneGiornaliera() const{return produzioneGiornaliera;}
